geolocation/requirements.txt
aniso8601==9.0.1
asyncio-nats-client==0.11.5
asyncio-nats-streaming==0.4.0
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
protobuf==3.20.1
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
pytz==2022.1
six==1.16.0
theartisans-shared-python==0.1.1
Werkzeug==2.1.2
zipp==3.8.0
